TransactionChainManager - OF plugin cluster fix
OFPlugin-Li was designed without cluster. To add clustering
corect support needs some design changes for TransactionChain.
We can not ignore relationship between TransactionChain and
Cluster role election. Because only MASTER could hold
TransactionChain to distributed DataStore. So we have to
redesign a TransactionChain lifecycle following the ownership
election results.

* add TransactionChainManagerStatus#SLEEPING for SLAVE
* add activation/deactivation TransactionChainManager methods
* mark unecessary code with deprecation marker
* fix tests
* reordered methods in order to simplify review

pathset 5:
fixed enabledSubmit boolean + test

TODOs:
(critical)
* add blocking functionalty to wait for Cluster election result
in RoleContext creation (DeviceNode Entity registration)
* CandidateAlreadyRegisteredException by registration has to
close actual postHanshake process
* add fork postHandshake steps for MASTER/SLAVE
* add additional WriteTransaction for last DisconnectedNode
Cluster holder
(no critical)
* clean code
* fix tests
* add new set of tests

+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/device/TransactionChainManager.java
@@ -14,9 +14,11 @@ import com.google.common.util.concurrent.CheckedFuture;
 import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
 import javax.annotation.Nonnull;
+import javax.annotation.concurrent.GuardedBy;
 import org.opendaylight.controller.md.sal.binding.api.BindingTransactionChain;
 import org.opendaylight.controller.md.sal.binding.api.DataBroker;
 import org.opendaylight.controller.md.sal.binding.api.WriteTransaction;
+import org.opendaylight.controller.md.sal.common.api.clustering.EntityOwnershipChange;
 import org.opendaylight.controller.md.sal.common.api.data.AsyncTransaction;
 import org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
 import org.opendaylight.controller.md.sal.common.api.data.TransactionChain;
@@ -51,39 +53,100 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
     private final Object txLock = new Object();
 
     private final DataBroker dataBroker;
+    @GuardedBy("txLock")
     private WriteTransaction wTx;
+    @GuardedBy("txLock")
     private BindingTransactionChain txChainFactory;
-    private boolean submitIsEnabled;
+    @Deprecated
+    private boolean submitIsEnabled = false;
 
-    public TransactionChainManagerStatus getTransactionChainManagerStatus() {
-        return transactionChainManagerStatus;
-    }
-
-    private TransactionChainManagerStatus transactionChainManagerStatus;
+    private volatile TransactionChainManagerStatus transactionChainManagerStatus;
+    @Deprecated
     private ReadyForNewTransactionChainHandler readyForNewTransactionChainHandler;
+
     private final KeyedInstanceIdentifier<Node, NodeKey> nodeII;
+    @Deprecated
     private volatile Registration managerRegistration;
 
     TransactionChainManager(@Nonnull final DataBroker dataBroker,
-                            @Nonnull final KeyedInstanceIdentifier<Node, NodeKey> nodeII,
-                            @Nonnull final Registration managerRegistration) {
+            @Nonnull final KeyedInstanceIdentifier<Node, NodeKey> nodeII,
+            @Nonnull final Registration managerRegistration) {
+        LOG.debug("TxChainManager initialization");
         this.dataBroker = Preconditions.checkNotNull(dataBroker);
         this.nodeII = Preconditions.checkNotNull(nodeII);
         this.managerRegistration = Preconditions.checkNotNull(managerRegistration);
-        this.transactionChainManagerStatus = TransactionChainManagerStatus.WORKING;
-        createTxChain(dataBroker);
-        LOG.debug("created txChainManager");
+        this.transactionChainManagerStatus = TransactionChainManagerStatus.SLEEPING;
+        LOG.trace("Initialization of txChainManager done");
     }
 
-    private void createTxChain(final DataBroker dataBroker) {
-        txChainFactory = dataBroker.createTransactionChain(TransactionChainManager.this);
+    /**
+     * Method return actual transactionChainManagerStatus. Method is used in DeviceManager to check
+     * how it could continue in postHanshake process. {@link TransactionChainManagerStatus#SLEEPING} means
+     * "we are SLAVE, so we can finish registration now" {@link TransactionChainManagerStatus#WORKING} means
+     * "we are MASTER, so we have to finish whole registration"
+     * @return actual {@link TransactionChainManagerStatus}
+     */
+    TransactionChainManagerStatus getTransactionChainManagerStatus() {
+        return transactionChainManagerStatus;
     }
 
+    /**
+     * Method change status for TxChainManager to {@link TransactionChainManagerStatus#WORKING} and it has to make
+     * registration for this class instance as {@link TransactionChainListener} to provide possibility a make DS
+     * transactions.
+     * @param ownershipChange - marker to be sure it is used only for MASTER
+     */
+    public void activateTransactionManager(final EntityOwnershipChange ownershipChange) {
+        LOG.trace("Changing txChain manager status to WORKING");
+        Preconditions.checkArgument(ownershipChange != null);
+        Preconditions.checkState(ownershipChange.isOwner());
+        if ((!ownershipChange.wasOwner()) && ownershipChange.isOwner()) {
+            synchronized (txLock) {
+                LOG.debug("Transaction Factory create");
+                Preconditions.checkState(txChainFactory == null, "TxChainFactory survive last close.");
+                Preconditions.checkState(wTx == null, "We have some unexpected WriteTransaction.");
+                this.transactionChainManagerStatus = TransactionChainManagerStatus.WORKING;
+                this.txChainFactory = dataBroker.createTransactionChain(TransactionChainManager.this);
+            }
+        } else {
+            LOG.debug("Ownership was not realy changed!");
+        }
+    }
+
+    /**
+     * Method change status for TxChainManger to {@link TransactionChainManagerStatus#SLEEPING} and it unregisters
+     * this class instance as {@link TransactionChainListener} so it broke a possibility to write something to DS
+     * @param ownershipChange - marker to be sure it is used only for SLAVE
+     */
+    public void deactivateTransactionManager(final EntityOwnershipChange ownershipChange) {
+        LOG.trace("Changing txChain manager status to SLEEPING");
+        Preconditions.checkArgument(ownershipChange != null);
+        Preconditions.checkState((!ownershipChange.isOwner()));
+        if (ownershipChange.wasOwner() && (!ownershipChange.isOwner())) {
+            synchronized (txLock) {
+                if (TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
+                    LOG.debug("Submitting all transactions if we were in status WORKING");
+                    submitWriteTransaction();
+                }
+                Preconditions.checkState(wTx == null, "We have some unexpected WriteTransaction.");
+                LOG.debug("Transaction Factory delete");
+                transactionChainManagerStatus = TransactionChainManagerStatus.SLEEPING;
+                txChainFactory.close();
+                txChainFactory = null;
+                wTx = null;
+            }
+        } else {
+            LOG.debug("Ownership was not realy changed!");
+        }
+    }
+
+    @Deprecated
     void initialSubmitWriteTransaction() {
         enableSubmit();
         submitWriteTransaction();
     }
 
+    @Deprecated
     public synchronized boolean attemptToRegisterHandler(final ReadyForNewTransactionChainHandler readyForNewTransactionChainHandler) {
         if (TransactionChainManagerStatus.SHUTTING_DOWN.equals(this.transactionChainManagerStatus)
                 && null == this.readyForNewTransactionChainHandler) {
@@ -97,12 +160,20 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
         }
     }
 
+    /**
+     * Method submits actual {@link WriteTransaction} to DataStore if some is available
+     *
+     * @return information about run submit process
+     */
     boolean submitWriteTransaction() {
-        if (!submitIsEnabled) {
+        LOG.trace("SubmitWriteTransaction Method");
+        if (!submitIsEnabled) { // this check will not need anymore (Status has to be working before call this method)
             LOG.trace("transaction not committed - submit block issued");
             return false;
         }
         synchronized (txLock) {
+            Preconditions.checkState(TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus),
+                    "we have here Uncompleted Transaction for node {} and we are not MASTER", nodeII);
             if (wTx == null) {
                 LOG.trace("nothing to commit - submit returns true");
                 return true;
@@ -110,16 +181,16 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
             final CheckedFuture<Void, TransactionCommitFailedException> submitFuture = wTx.submit();
             Futures.addCallback(submitFuture, new FutureCallback<Void>() {
                 @Override
-                public void onSuccess(Void result) {
+                public void onSuccess(final Void result) {
                     //no action required
                 }
 
                 @Override
-                public void onFailure(Throwable t) {
+                public void onFailure(final Throwable t) {
                     if (t instanceof TransactionCommitFailedException) {
-                        LOG.error("Transaction commit failed. {}", t);
+                        LOG.error("Transaction for node {} commit failed. {}", nodeII, t);
                     } else {
-                        LOG.error("Exception during transaction submitting. {}", t);
+                        LOG.error("Exception during transaction submitting for node {}. {}", nodeII, t);
                     }
                 }
             });
@@ -128,49 +199,100 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
         return true;
     }
 
+    @Deprecated
     public void cancelWriteTransaction() {
         // there is no cancel txn in ping-pong broker. So we need to drop the chain and recreate it.
         // since the chain is created per device, there won't be any other txns other than ones we created.
         recreateTxChain();
     }
 
+    /**
+     * Method deletes object identified by PATH from DS
+     * @param store Operational or Configuration
+     * @param path the Deleted object identifier
+     */
     <T extends DataObject> void addDeleteOperationTotTxChain(final LogicalDatastoreType store,
                                                              final InstanceIdentifier<T> path) {
-        final WriteTransaction writeTx = getTransactionSafely();
-        writeTx.delete(store, path);
+        LOG.trace("addDeleteOperationToTxChain method call");
+        Preconditions.checkArgument(store != null);
+        Preconditions.checkArgument(path != null);
+        if (TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
+            LOG.debug("Deleting {} from {} DataStore.", path, store);
+            final WriteTransaction writeTx = getTransactionSafely();
+            if (writeTx != null) {
+                writeTx.delete(store, path);
+            } else {
+                LOG.info("You are not MASTER and Delete {} was not writen to {} DS", path, store);
+            }
+        } else {
+            LOG.warn("You try to delete {} from {} DataStore, but you are not MASTER", path, store);
+        }
     }
 
+    /**
+     * Method writes data identified by PATH from DS
+     * @param store Operational or Configuration
+     * @param path the new data identifier
+     * @param data dataObject to add
+     */
     <T extends DataObject> void writeToTransaction(final LogicalDatastoreType store,
                                                    final InstanceIdentifier<T> path, final T data) {
-        final WriteTransaction writeTx = getTransactionSafely();
-        writeTx.put(store, path, data);
+        LOG.trace("writeToTransaction method call");
+        Preconditions.checkArgument(store != null);
+        Preconditions.checkArgument(path != null);
+        Preconditions.checkArgument(data != null);
+        if (TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
+            LOG.debug("Writing data {} to {} DataStore.", path, store);
+            final WriteTransaction writeTx = getTransactionSafely();
+            if (writeTx != null) {
+                writeTx.put(store, path, data);
+            } else {
+                LOG.info("You are not MASTER and Write data to {} was not writen to {} DS", path, store);
+            }
+        } else {
+            LOG.warn("You try to write data {} to {} DataStore, but you are not MASTER", path, store);
+        }
     }
 
     @Override
     public void onTransactionChainFailed(final TransactionChain<?, ?> chain,
                                          final AsyncTransaction<?, ?> transaction, final Throwable cause) {
-        LOG.warn("txChain failed -> recreating", cause);
-        recreateTxChain();
+        LOG.warn("Submited TxChain for node {} Failed -> recreating", nodeII, cause);
+        synchronized (txLock) {
+            if (TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
+                this.txChainFactory.close();
+                this.txChainFactory = dataBroker.createTransactionChain(TransactionChainManager.this);
+                this.wTx = null;
+            }
+        }
     }
 
     @Override
     public void onTransactionChainSuccessful(final TransactionChain<?, ?> chain) {
+        LOG.trace("Submited TxChain for node {} finish Successful", nodeII);
         // NOOP - only yet, here is probably place for notification to get new WriteTransaction
     }
 
+    @Deprecated
     private void recreateTxChain() {
-        txChainFactory.close();
-        createTxChain(dataBroker);
         synchronized (txLock) {
-            wTx = null;
+            this.txChainFactory.close();
+            this.txChainFactory = dataBroker.createTransactionChain(TransactionChainManager.this);
+            this.wTx = null;
         }
     }
 
-
+    /**
+     * Method provide {@link WriteTransaction} to write or delete data to DataStore
+     *
+     * @return actual or new transaction
+     */
     private WriteTransaction getTransactionSafely() {
-        if (wTx == null && !TransactionChainManagerStatus.SHUTTING_DOWN.equals(transactionChainManagerStatus)) {
+        LOG.trace("getTransactionSafely method");
+        if (wTx == null && TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
             synchronized (txLock) {
-                if (wTx == null) {
+                if (wTx == null && TransactionChainManagerStatus.WORKING.equals(transactionChainManagerStatus)) {
+                    Preconditions.checkState(txChainFactory != null);
                     wTx = txChainFactory.newWriteOnlyTransaction();
                 }
             }
@@ -179,6 +301,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
     }
 
     @VisibleForTesting
+    @Deprecated
     void enableSubmit() {
         submitIsEnabled = true;
     }
@@ -190,6 +313,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
      * had disconnected, then we clean the inventory.
      * Called from DeviceContext
      */
+    @Deprecated
     public void cleanupPostClosure() {
         LOG.debug("Removing node {} from operational DS.", nodeII);
         synchronized (txLock) {
@@ -199,7 +323,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
             if (txChainFactory == null) {
                 LOG.info("Creating new Txn Chain Factory for cleanup purposes - Race Condition Hazard, " +
                         "Concurrent Modification Hazard, node:{}", nodeII);
-                createTxChain(dataBroker);
+                this.txChainFactory = dataBroker.createTransactionChain(TransactionChainManager.this);
             }
 
             if (TransactionChainManagerStatus.SHUTTING_DOWN.equals(transactionChainManagerStatus)) {
@@ -213,7 +337,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
             writeTx.delete(LogicalDatastoreType.OPERATIONAL, nodeII);
             LOG.debug("Delete node {} from operational DS put to write transaction.", nodeII);
 
-            CheckedFuture<Void, TransactionCommitFailedException> submitsFuture = writeTx.submit();
+            final CheckedFuture<Void, TransactionCommitFailedException> submitsFuture = writeTx.submit();
             LOG.debug("Delete node {} from operational DS write transaction submitted.", nodeII);
 
             Futures.addCallback(submitsFuture, new FutureCallback<Void>() {
@@ -233,6 +357,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
         }
     }
 
+    @Deprecated
     private void notifyReadyForNewTransactionChainAndCloseFactory() {
         if(managerRegistration == null){
             LOG.warn("managerRegistration is null");
@@ -244,7 +369,7 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
                     LOG.debug("Closing registration in manager.");
                     managerRegistration.close();
                 }
-            } catch (Exception e) {
+            } catch (final Exception e) {
                 LOG.warn("Failed to close transaction chain manager's registration.", e);
             }
             managerRegistration = null;
@@ -252,23 +377,30 @@ class TransactionChainManager implements TransactionChainListener, AutoCloseable
                 readyForNewTransactionChainHandler.onReadyForNewTransactionChain();
             }
         }
-        txChainFactory.close();
-        txChainFactory = null;
+        if (txChainFactory != null) {
+            txChainFactory.close();
+            txChainFactory = null;
+        }
         LOG.debug("Transaction chain factory closed.");
     }
 
     @Override
     public void close() {
         LOG.debug("closing txChainManager without cleanup of node {} from operational DS.", nodeII);
+        Preconditions.checkState(wTx == null);
+        Preconditions.checkState(txChainFactory == null);
         synchronized (txLock) {
             this.transactionChainManagerStatus = TransactionChainManagerStatus.SHUTTING_DOWN;
-            notifyReadyForNewTransactionChainAndCloseFactory();
-            wTx = null;
         }
+        notifyReadyForNewTransactionChainAndCloseFactory(); // don't need anymore
     }
 
     public enum TransactionChainManagerStatus {
-        WORKING, SHUTTING_DOWN;
+        /** txChainManager is sleeping - is not active (SLAVE or default init value) */
+        SLEEPING,
+        /** txChainManager is working - is active (MASTER) */
+        WORKING,
+        /** txChainManager is trying to be closed - device disconnecting */
+        SHUTTING_DOWN;
     }
-
 }
